--- src/database/seeds/BantenProvJenisPrestasiSeeder.csv
@@ -1,3 +1,0 @@
-jenis Prestasi 1,1
-jenis prestasi 2,2
-jenis prestasi 3,3
