--- src/database/seeds/BantenprovPrestasiSeeder.csv
@@ -1,2 +1,0 @@
-1,1,1,lomba 1
-2,2,2,lomba 2
