+++ src/database/seeds/BantenprovPrestasiSeeder.php
@@ -1,32 +1,9 @@
 <?php
+
 use Illuminate\Database\Seeder;
-/**
- * Usage :
- * [1] $ composer dump-autoload -o
- * [2] $ php artisan db:seed --class=BantenprovPrestasiSeeder
- */
+
 class BantenprovPrestasiSeeder extends Seeder
 {
-    /* text color */
-    protected $RED     ="\033[0;31m";
-    protected $CYAN    ="\033[0;36m";
-    protected $YELLOW  ="\033[1;33m";
-    protected $ORANGE  ="\033[0;33m";
-    protected $PUR     ="\033[0;35m";
-    protected $GRN     ="\e[32m";
-    protected $WHI     ="\e[37m";
-    protected $NC      ="\033[0m";
-    /* File name */
-    /* location : /databse/seeds/file_name.csv */
-    protected $fileName = "BantenprovPrestasiSeeder.csv";
-    /* text info : default (true) */
-    protected $textInfo = true;
-    /* model class */
-    protected $model;
-    /* __construct */
-    public function __construct(){
-        $this->model = new Bantenprov\Prestasi\Models\Bantenprov\Prestasi\Prestasi;
-    }
     /**
      * Run the database seeds.
      *
@@ -34,63 +11,8 @@ class BantenprovPrestasiSeeder extends Seeder
      */
     public function run()
     {
-        $this->insertData();
-    }
-    /* function insert data */
-    protected function insertData()
-    {
-        /* silahkan di rubah sesuai kebutuhan */
-        foreach($this->readCSV() as $data){
-
-            
-        	$this->model->create([
-            	'user_id' => $data['user_id'],
-				'master_prestasi_id' => $data['master_prestasi_id'],
-				'nomor_un' => $data['nomor_un'],
-				'nama_lomba' => $data['nama_lomba'],
-
-        	]);
-        
-
-        }
-
-        if($this->textInfo){                
-            echo "============[DATA]============\n";
-            $this->orangeText('user_id : ').$this->greenText($data['user_id']);
-			echo"\n";
-			$this->orangeText('master_prestasi_id : ').$this->greenText($data['master_prestasi_id']);
-			echo"\n";
-			$this->orangeText('nomor_un : ').$this->greenText($data['nomor_un']);
-			echo"\n";
-			$this->orangeText('nama_lomba : ').$this->greenText($data['nama_lomba']);
-			echo"\n";
-        
-            echo "============[DATA]============\n\n";
-        }
-
-        $this->greenText('[ SEEDER DONE ]');
-        echo"\n\n";
-    }
-    /* text color: orange */
-    protected function orangeText($text)
-    {
-        printf($this->ORANGE.$text.$this->NC);
-    }
-    /* text color: green */
-    protected function greenText($text)
-    {
-        printf($this->GRN.$text.$this->NC);
-    }
-    /* function read CSV file */
-    protected function readCSV()
-    {
-        $file = fopen(database_path("seeds/".$this->fileName), "r");
-        $all_data = array();
-        $row = 1;
-        while(($data = fgetcsv($file, 1000, ",")) !== FALSE){
-            $all_data[] = ['user_id' => $data[0],'master_prestasi_id' => $data[1],'nomor_un' => $data[2],'nama_lomba' => $data[3],];
-        }
-        fclose($file);
-        return  $all_data;
+        $this->call(BantenprovPrestasiSeederJenisPrestasi::class);
+        $this->call(BantenprovPrestasiSeederMasterPrestasi::class);
+        $this->call(BantenprovPrestasiSeederPrestasi::class);
     }
 }
