+++ src/database/seeds/BantenprovPrestasiSeederJenisPrestasi.csv
@@ -1,0 +1,3 @@
+Jenis Prestasi 1,1
+Jenis prestasi 2,2
+Jenis prestasi 3,3
