+++ src/database/seeds/BantenprovPrestasiSeederJenisPrestasi.php
@@ -3,9 +3,9 @@ use Illuminate\Database\Seeder;
 /**
  * Usage :
  * [1] $ composer dump-autoload -o
- * [2] $ php artisan db:seed --class=BantenprovMasterPrestasiSeeder
+ * [2] $ php artisan db:seed --class=BantenprovPrestasiSeederJenisPrestasi
  */
-class BantenprovMasterPrestasiSeeder extends Seeder
+class BantenprovPrestasiSeederJenisPrestasi extends Seeder
 {
     /* text color */
     protected $RED     ="\033[0;31m";
@@ -18,14 +18,14 @@ class BantenprovMasterPrestasiSeeder extends Seeder
     protected $NC      ="\033[0m";
     /* File name */
     /* location : /databse/seeds/file_name.csv */
-    protected $fileName = "BantenprovMasterPrestasiSeeder.csv";
+    protected $fileName = "BantenprovPrestasiSeederJenisPrestasi.csv";
     /* text info : default (true) */
     protected $textInfo = true;
     /* model class */
     protected $model;
     /* __construct */
     public function __construct(){
-        $this->model = new Bantenprov\Prestasi\Models\Bantenprov\Prestasi\MasterPrestasi;
+        $this->model = new Bantenprov\Prestasi\Models\Bantenprov\Prestasi\JenisPrestasi;
     }
     /**
      * Run the database seeds.
@@ -42,35 +42,23 @@ class BantenprovMasterPrestasiSeeder extends Seeder
         /* silahkan di rubah sesuai kebutuhan */
         foreach($this->readCSV() as $data){
 
-            
+
         	$this->model->create([
-            	'jenis_prestasi_id' => $data['jenis_prestasi_id'],
-				'juara' => $data['juara'],
-				'tingkat' => $data['tingkat'],
-				'nilai' => $data['nilai'],
-				'kode_prestasi' => $data['kode_prestasi'],
+            	'nama_jenis_prestasi' => $data['nama_jenis_prestasi'],
 				'user_id' => $data['user_id'],
 
         	]);
-        
+
 
         }
 
-        if($this->textInfo){                
+        if($this->textInfo){
             echo "============[DATA]============\n";
-            $this->orangeText('jenis_prestasi_id : ').$this->greenText($data['jenis_prestasi_id']);
-			echo"\n";
-			$this->orangeText('juara : ').$this->greenText($data['juara']);
-			echo"\n";
-			$this->orangeText('tingkat : ').$this->greenText($data['tingkat']);
-			echo"\n";
-			$this->orangeText('nilai : ').$this->greenText($data['nilai']);
-			echo"\n";
-			$this->orangeText('kode_prestasi : ').$this->greenText($data['kode_prestasi']);
+            $this->orangeText('nama_jenis_prestasi : ').$this->greenText($data['nama_jenis_prestasi']);
 			echo"\n";
 			$this->orangeText('user_id : ').$this->greenText($data['user_id']);
 			echo"\n";
-        
+
             echo "============[DATA]============\n\n";
         }
 
@@ -94,7 +82,10 @@ class BantenprovMasterPrestasiSeeder extends Seeder
         $all_data = array();
         $row = 1;
         while(($data = fgetcsv($file, 1000, ",")) !== FALSE){
-            $all_data[] = ['jenis_prestasi_id' => $data[0],'juara' => $data[1],'tingkat' => $data[2],'nilai' => $data[3],'kode_prestasi' => $data[4],'user_id' => $data[5],];
+            $all_data[] = [
+                'nama_jenis_prestasi' => $data[0],
+                'user_id' => $data[1],
+            ];
         }
         fclose($file);
         return  $all_data;
