+++ src/database/seeds/BantenprovPrestasiSeederPrestasi.csv
@@ -1,0 +1,2 @@
+1,1,Lomba Kerupuk,100,1
+2,2,Lomba Karung,50,1
